requirements.txt
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-telegram-bot==20.8
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.2
Werkzeug==3.0.3
